DemoSlectionActivity created. Rename MainActivity -> NetworkDemoActivity
