+++ sample/build.gradle
@@ -26,6 +26,7 @@ android {
 
     buildFeatures {
         dataBinding true
+        viewBinding true
     }
 }
