+++ sample/src/main/AndroidManifest.xml
@@ -13,14 +13,15 @@
         android:usesCleartextTraffic="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".MainActivity">
+        <activity android:name=".DemoSelectionActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name=".DetailsActivity" />
+        <activity android:name=".network.NetworkDemoActivity"/>
+        <activity android:name=".network.NetworkDemoDetailsActivity" />
 
     </application>
