+++ sample/src/main/java/com/daydreamapplications/materialdesignadapter/network/NetworkDemoDetailsActivity.kt
@@ -1,4 +1,4 @@
-package com.daydreamapplications.materialdesignadapter
+package com.daydreamapplications.materialdesignadapter.network
 
 import android.content.Context
 import android.content.Intent
@@ -9,17 +9,17 @@ import com.bumptech.glide.Glide
 import com.daydreamapplications.adapter.TransitionNames
 import com.daydreamapplications.materialdesignadapter.databinding.ActivityDetailsBinding
 
-class DetailsActivity : AppCompatActivity() {
+class NetworkDemoDetailsActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val foodItem = intent.getParcelableExtra<FoodItem>("foodItem") ?: throw IllegalStateException("FoodItem data missing")
+        val foodItem = intent.getParcelableExtra<NetworkDataProvider.FoodItem>("foodItem") ?: throw IllegalStateException("FoodItem data missing")
 
         ActivityDetailsBinding.inflate(layoutInflater)
             .apply {
 
                 banner.transitionName = TransitionNames.visual
-                Glide.with(this@DetailsActivity)
+                Glide.with(this@NetworkDemoDetailsActivity)
                     .load(foodItem.imageUrl)
                     .into(banner)
 
@@ -33,12 +33,12 @@ class DetailsActivity : AppCompatActivity() {
 
     companion object {
 
-        fun startActivity(context: Context, foodItem: FoodItem, options: Bundle? = null) {
+        fun startActivity(context: Context, foodItem: NetworkDataProvider.FoodItem, options: Bundle? = null) {
             ActivityCompat.startActivity(context, startActivityIntent(context, foodItem), options)
         }
 
-        fun startActivityIntent(context: Context, foodItem: FoodItem): Intent {
-            return Intent(context, DetailsActivity::class.java)
+        fun startActivityIntent(context: Context, foodItem: NetworkDataProvider.FoodItem): Intent {
+            return Intent(context, NetworkDemoDetailsActivity::class.java)
                 .putExtra("foodItem", foodItem)
         }
     }
