+++ sample/src/main/res/layout/activity_demo_selection.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:orientation="vertical"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:gravity="center"
+        android:text="Further Examples" />
+
+    <Button
+        android:id="@+id/button_network_example"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Live Network Example" />
+
+</LinearLayout>
